Extact class - Split CommandHandler to  CommandHandlerParseLocation, CommandHandlerParseTime, CommandHandlerParseWorld
