+++ src/main/java/net/coreprotect/bukkit/BukkitInterface.java
@@ -43,5 +43,19 @@ public interface BukkitInterface {
     public void sendSignChange(Player player, Sign sign);
 
     public void setGlowing(Sign sign, boolean b);
+    public String parseLegacyName(String name);
+
+    int getLegacyBlockId(Material material);
+
+    public boolean getEntityMeta(LivingEntity entity, List<Object> info);
+
+    public boolean setEntityMeta(Entity entity, Object value, int count);
+
+    public boolean getItemMeta(ItemMeta itemMeta, List<Map<String, Object>> list, List<List<Map<String, Object>>> metadata, int slot);
+
+    public boolean setItemMeta(Material rowType, ItemStack itemstack, List<Map<String, Object>> map);
+
+    public boolean isAttached(Block block, Block scanBlock, BlockData blockData, int scanMin);
+
 
 }
