+++ src/main/java/net/coreprotect/command/LookupCommand.java
@@ -43,24 +43,24 @@ public class LookupCommand {
     protected static void runCommand(CommandSender player, Command command, boolean permission, String[] args) {
         int resultc = args.length;
         args = CommandHandler.parsePage(args);
-        Location lo = CommandHandler.parseLocation(player, args);
+        Location lo = CommandHandlerParseLocation.parseLocation(player, args);
         // List<String> arg_uuids = new ArrayList<String>();
         List<String> argUsers = CommandHandler.parseUsers(args);
-        Integer[] argRadius = CommandHandler.parseRadius(args, player, lo);
+        Integer[] argRadius = CommandHandlerParseLocation.parseRadius(args, player, lo);
         int argNoisy = CommandHandler.parseNoisy(args);
         List<Integer> argAction = CommandHandler.parseAction(args);
-        List<Object> argBlocks = CommandHandler.parseRestricted(player, args, argAction);
+        List<Object> argBlocks = CommandHandlerParseLocation.parseRestricted(player, args, argAction);
         Map<Object, Boolean> argExclude = CommandHandler.parseExcluded(player, args, argAction);
         List<String> argExcludeUsers = CommandHandler.parseExcludedUsers(player, args);
-        String ts = CommandHandler.parseTimeString(args);
-        long[] argTime = CommandHandler.parseTime(args);
+        String ts = CommandHandlerParseLocation.parseTimeString(args);
+        long[] argTime = CommandHandlerParseLocation.parseTime(args);
         long startTime = argTime[0];
         long endTime = argTime[1];
-        int argWid = CommandHandler.parseWorld(args, true, true);
+        int argWid = CommandHandlerParseWorld.parseWorld(args, true, true);
         int parseRows = CommandHandler.parseRows(args);
         boolean count = CommandHandler.parseCount(args);
-        boolean worldedit = CommandHandler.parseWorldEdit(args);
-        boolean forceglobal = CommandHandler.parseForceGlobal(args);
+        boolean worldedit = CommandHandlerParseWorld.parseWorldEdit(args);
+        boolean forceglobal = CommandHandlerParseLocation.parseForceGlobal(args);
         boolean pageLookup = false;
 
         if (argBlocks == null || argExclude == null || argExcludeUsers == null) {
@@ -117,7 +117,7 @@ public class LookupCommand {
         }
 
         if (argWid == -1) {
-            String worldName = CommandHandler.parseWorldName(args, true);
+            String worldName = CommandHandlerParseWorld.parseWorldName(args, true);
             Chat.sendMessage(player, new ChatMessage(Phrase.build(Phrase.WORLD_NOT_FOUND, worldName)).build());
             return;
         }
